requirements.txt
opencv-python
numpy
mediapipe
Pillow
